move logic to hooks
